requirements.txt
# Core
easyocr>=1.7.0
sentence-transformers>=2.2.2
networkx>=3.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0

# Web UI
streamlit>=1.28.0

# Optional interactive graph
pyvis>=0.3.2

# COCO (annotation parsing — optional, we use stdlib json directly)
# pycocotools>=2.0.6   # uncomment if you need mask/polygon support
